Create README.md
This code was given as part of our course project. Our task was to make it more secure and detected the errors in the code. These are the things we required to fix 1) Logical and lexical Errors 2) Class, Variable and Argument hardening 3) Exception handling  4) Redundant Null checking 5) FindBug Analysis to check the code for bug in Java code.